saftbus-gen finds all bases classes

+++ example-plugin/src/Dice.cpp
@@ -6,7 +6,7 @@ Dice::Dice() {
 	max = 6;
 }
 
-int Dice::roll() {
+double Dice::roll() {
 	return 1+rand()%max;
 }
